requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
celery==5.3.6
redis==5.0.1
django-filter==23.5
drf-yasg==1.21.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
setuptools==78.1.0